🐛fix(clickmouse):修复了init无法运行的bug
- [🐛fix]修复了init无法运行的bug
- [❌remove]移除了无用的资源
- [❌remove]移除了官方版本校验
- [📆chore]修改了makefile，移除了无用内容

+++ Gui/install_pack.py
@@ -845,7 +845,8 @@ if __name__ == '__main__':
         for i in package_name:
             packages_source.append(import_package(i))
     except FileNotFoundError:
-        os.remove(Path('data', 'first_run'))
+        if os.path.exists('data/first_run'):
+            os.remove(Path('data', 'first_run'))
     
     packages = get_packages()
